fix: fixed issue with units for dollar values
After recomparing the value of work completed to knowledge of the industry caused me to question the accuracy of the value of work completed. For example I knew that in 2009-2010 the Australian Government invested $14b into the Building Education Revolution project. I rechecked the source data and could see that the units were $000

Updated prepData.py and graphs